+++ README.md
@@ -1,6 +1,6 @@
 # BlenderNC
 
-![Read the Docs](https://img.shields.io/readthedocs/blendernc) [![Maintainability](https://api.codeclimate.com/v1/badges/bbd6f981e5f5a26c6a56/maintainability)](https://codeclimate.com/github/blendernc/blendernc/maintainability) [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black) ![GitHub](https://img.shields.io/github/license/blendernc/blendernc?color=lightblue) ![blender support](https://img.shields.io/badge/blender-2.83--2.92-blueviolet) ![GitHub release (latest by date)](https://img.shields.io/github/v/release/blendernc/blendernc?label=tag) [![pre-commit.ci status](https://results.pre-commit.ci/badge/github/blendernc/blendernc/master.svg)](https://results.pre-commit.ci/latest/github/blendernc/blendernc/master)
+![Read the Docs](https://img.shields.io/readthedocs/blendernc) ![Github CI](https://github.com/blendernc/blendernc/actions/workflows/test.yml/badge.svg) [![Maintainability](https://api.codeclimate.com/v1/badges/bbd6f981e5f5a26c6a56/maintainability)](https://codeclimate.com/github/blendernc/blendernc/maintainability) [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black) ![GitHub](https://img.shields.io/github/license/blendernc/blendernc?color=lightblue) ![blender support](https://img.shields.io/badge/blender-2.83--2.92-blueviolet) ![GitHub release (latest by date)](https://img.shields.io/github/v/release/blendernc/blendernc?label=tag) [![pre-commit.ci status](https://results.pre-commit.ci/badge/github/blendernc/blendernc/master.svg)](https://results.pre-commit.ci/latest/github/blendernc/blendernc/master)
 
 **BlenderNC** is an open source add-on and Python module to visualize **netCDF** data in [**Blender**](www.blender.org).
 
@@ -11,9 +11,11 @@ BlenderNC builds upon [**xarray**](https://github.com/pydata/xarray) and [**dask
 Science visualization is a fundamental part of science communication and exploration of large datasets. However, production quality real-time visualization and animation of scientific data has remain unreachable to the scientific community. BlenderNC main goal is to facilitate generation of quality animations of scientific gridded data with a powerful and simple interface. For example:
 
 - Quick load of netCDFs:
+
 <img src="./docs/images/quick_load_gif.gif" width="70%" />
 
 - Nodes tree for more complex visualizations:
+
 <img src="./docs/images/GEBCO_blendernc.png" width="70%" />
 
 - Math computations in BlenderNC node tree.
