+++ blendernc/nodes/grid/BlenderNC_NT_input_grid.py
@@ -4,8 +4,8 @@ from collections import defaultdict
 
 import bpy
 
-from ....blendernc.decorators import NodesDecorators
-from ....blendernc.python_functions import get_new_identifier, get_var
+from blendernc.decorators import NodesDecorators
+from blendernc.python_functions import get_new_identifier, get_var
 
 
 def get_possible_grid(node, context):
