+++ blendernc/nodes/grid/BlenderNC_NT_rotate_lon.py
@@ -4,8 +4,8 @@ from collections import defaultdict
 
 import bpy
 
-from ....blendernc.decorators import NodesDecorators
-from ....blendernc.python_functions import rotate_longitude, update_value_and_node_tree
+from blendernc.decorators import NodesDecorators
+from blendernc.python_functions import rotate_longitude, update_value_and_node_tree
 
 
 class BlenderNC_NT_rotatelon(bpy.types.Node):
