+++ blendernc/nodes/outputs/BlenderNC_NT_output.py
@@ -4,9 +4,9 @@ from collections import defaultdict
 
 import bpy
 
-from ....blendernc.decorators import NodesDecorators
-from ....blendernc.image import dataset_2_image_preview
-from ....blendernc.python_functions import (
+from blendernc.decorators import NodesDecorators
+from blendernc.image import dataset_2_image_preview
+from blendernc.python_functions import (
     update_colormap_interface,
     update_image,
     update_value,
