+++ blendernc/nodes/outputs/BlenderNC_NT_preloader.py
@@ -2,7 +2,7 @@
 # Imports
 import bpy
 
-from ....blendernc.python_functions import get_possible_files, get_possible_variables
+from blendernc.python_functions import get_possible_files, get_possible_variables
 
 
 class BlenderNC_NT_preloader(bpy.types.Node):
