+++ blendernc/nodes/selecting/BlenderNC_NT_drop_dims.py
@@ -4,8 +4,8 @@ from collections import defaultdict
 
 import bpy
 
-from ....blendernc.decorators import NodesDecorators
-from ....blendernc.python_functions import get_possible_dims, update_value_and_node_tree
+from blendernc.decorators import NodesDecorators
+from blendernc.python_functions import get_possible_dims, update_value_and_node_tree
 
 
 class BlenderNC_NT_drop_dims(bpy.types.Node):
