+++ blendernc/nodes/selecting/BlenderNC_NT_select_axis.py
@@ -4,9 +4,9 @@ from collections import defaultdict
 
 import bpy
 
-from ....blendernc.core.netcdf_metadata import get_items_axes
-from ....blendernc.decorators import NodesDecorators
-from ....blendernc.python_functions import (
+from blendernc.core.netcdf_metadata import get_items_axes
+from blendernc.decorators import NodesDecorators
+from blendernc.python_functions import (
     refresh_cache,
     update_node_tree,
     update_value_and_node_tree,
