+++ setup.py
@@ -35,9 +35,10 @@ setup(
     author="josuemtzmo",
     author_email="josue.martinezmoreno@anu.edu.au",
     license="MIT License",
-    packages=[],
+    package_dir={"blendernc": "./blendernc"},
+    packages=["blendernc"],
     install_requires=["cython", "numpy"],
-    zip_safe=True,
+    zip_safe=False,
     # cmdclass={"build_ext": CustomBuildExtCommand},
     # TODO: Fix issue with blender not having python headers.
     # ext_modules=cythonize([extensions], build_dir="cython_build"),
